# internal/infrastructure/http/types/response_types.go
package types

// ErrorResponse represents a standardized error response.
type ErrorResponse struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
	Details string `json:"details,omitempty"`
}